# Conflicts:
#	src/client/java/lpctools/lpcfymasaapi/render/translucentShapes/LineCube.java
#	src/client/java/lpctools/lpcfymasaapi/render/translucentShapes/LineQuad.java
#	src/client/java/lpctools/lpcfymasaapi/render/translucentShapes/RenderInstance.java
#	src/client/java/lpctools/tools/canSpawnDisplay/IRenderMethod.java
#	src/client/java/lpctools/tools/canSpawnDisplay/LineCubeRenderMethod.java
#	src/client/java/lpctools/tools/canSpawnDisplay/MinihudStyleRenderMethod.java
#	src/client/java/lpctools/tools/furnaceMaintainer/FurnaceMaintainerRunner.java

+++ src/client/java/lpctools/lpcfymasaapi/configButtons/uniqueConfigs/UniqueDoubleConfig.java
@@ -32,7 +32,7 @@ public class UniqueDoubleConfig extends LPCUniqueConfigBase implements IConfigDo
         this.maxDouble = maxDouble;
     }
     @SuppressWarnings("unused")
-	public UniqueDoubleConfig(@NotNull ILPCConfigReadable parent, @NotNull String nameKey, int defaultDouble) {
+	public UniqueDoubleConfig(@NotNull ILPCConfigReadable parent, @NotNull String nameKey, double defaultDouble) {
         this(parent, nameKey, defaultDouble, -Double.MAX_VALUE, Double.MAX_VALUE, null);
     }
     @Override public double getDoubleValue() {return doubleValue;}
